+++ TurnkeyReact/components/Oauth.tsx
@@ -1,12 +1,12 @@
 import { useTurnkey } from "@turnkey/sdk-react-native";
 import { useEffect, useState, useCallback } from "react";
-import { TurnkeyConfigs } from "./SharedConfigs";
+import { TurnkeyConfigs } from "../SharedConfigs";
 import { Button } from "./ui/button";
 import { View, Text } from "react-native";
 // import GoogleIcon from "../assets/svgs/google.svg";
-import { OAUTH_TOKEN_EXPIRATION_SECONDS } from "./lib/constants";
+import { OAUTH_TOKEN_EXPIRATION_SECONDS } from "../lib/constants";
 // import * as Crypto from "expo-crypto";
-import { styles } from "./TurnkeyStyle";
+import { styles } from "../turnkeyStyle";
 import SHA256 from 'crypto-js/sha256';
 
 type OAuthProps =  {
