+++ TurnkeyReact/components/ui/button.tsx
@@ -2,7 +2,7 @@ import { cva, type VariantProps } from "class-variance-authority";
 import * as React from "react";
 import { Pressable } from "react-native";
 import { TextClassContext } from "./text";
-import { cn } from "../lib/utils";
+import { cn } from "../../lib/utils";
 import { Spinner } from "./spinner";
 
 const buttonVariants = cva(
