+++ TurnkeyReact/components/ui/spinner.tsx
@@ -7,7 +7,7 @@ import Animated, {
   withTiming,
   Easing,
 } from "react-native-reanimated";
-import { cn } from "../lib/utils";
+import { cn } from "../../lib/utils";
 import { Loader } from "lucide-react-native";
 
 const Spinner = React.forwardRef<View, { className?: string; size?: number }>(
