+++ TurnkeyReact/components/ui/text.tsx
@@ -2,7 +2,7 @@ import * as Slot from "@rn-primitives/slot";
 import type { SlottableTextProps, TextRef } from "@rn-primitives/types";
 import * as React from "react";
 import { Text as RNText } from "react-native";
-import { cn } from "../lib/utils";
+import { cn } from "../../lib/utils";
 
 const TextClassContext = React.createContext<string | undefined>(undefined);
