+++ package.json
@@ -30,6 +30,7 @@
     "react-native": "0.78.2",
     "react-native-crypto": "^2.2.1",
     "react-native-css-interop": "^0.1.22",
+    "react-native-gesture-handler": "^2.27.2",
     "react-native-get-random-values": "^1.11.0",
     "react-native-inappbrowser-reborn": "^3.7.0",
     "react-native-keychain": "^9.2.3",
@@ -47,6 +48,7 @@
     "@babel/core": "^7.25.2",
     "@babel/preset-env": "^7.25.3",
     "@babel/runtime": "^7.25.0",
+    "@eslint/js": "^9.31.0",
     "@react-native-community/cli": "15.0.1",
     "@react-native-community/cli-platform-android": "15.0.1",
     "@react-native-community/cli-platform-ios": "15.0.1",
@@ -59,11 +61,14 @@
     "@types/react": "^19.0.0",
     "@types/react-native-vector-icons": "^6.4.18",
     "@types/react-test-renderer": "^19.0.0",
-    "eslint": "^8.19.0",
+    "eslint": "^8.57.1",
+    "eslint-plugin-react": "^7.37.5",
+    "globals": "^16.3.0",
     "jest": "^29.6.3",
     "prettier": "2.8.8",
     "react-test-renderer": "19.0.0",
-    "typescript": "5.0.4"
+    "typescript": "5.0.4",
+    "typescript-eslint": "^8.38.0"
   },
   "engines": {
     "node": ">=18"
